requirements.txt
psycopg2==2.9.3
python-dateutil~=2.8.2
bs4~=0.0.1
lxml
py-cord
py-cord[voice]
yt-dlp
aiohttp~=3.7.4.post0
asyncio~=3.4.3
beautifulsoup4~=4.10.0
spotipy~=2.19.0
yarl==1.7.2